Split up creation of individual ingredient so identifier can be returned
